moved rowIndexes calculation logic
