+++ styles/themes.less
@@ -365,6 +365,10 @@
     line-height: 40px;
   }
 
+  .groups-table-row {
+       padding-top: 40px;
+  }
+
   .page {
     white-space: normal;
 
@@ -539,6 +543,10 @@
     font-weight: bold;
     opacity: 1;
     font-size: 17px;
+  }
+
+  .groups-table-row {
+    padding-top: 40px;
   }
 
   .weather-list table tr:nth-child(2n) {
